pyproject.toml
[project]
name = "meme-hunter"
version = "0.1.0"
description = "梗图猎人 - 抓全网热点用 Wan2.7 自动生成梗图"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "pillow>=10.0",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "dashscope>=1.20",
    "fastapi>=0.115",
    "uvicorn>=0.30",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["meme_hunter*"]
